wayland: Resturcture event handling.
We were using wl_display_roundtrip() in our run loop which was working but it
was wrong. Changing to wl_display_dispatch_pending() wasn't enough to keep
proper event handling functionality. This patch reworks the event loop so
timers work as expected and events are continuously delivered. Some of this
code was borrowed from weston.